Перевод тестов на @testing-library/react
