+++ src/Client/package.json
@@ -91,7 +91,6 @@
     "postcss": "^8.5.6",
     "postcss-loader": "^8.2.0",
     "prettier": "^3.6.2",
-    "react-test-renderer": "^18.3.1",
     "sass": "^1.93.2",
     "sass-loader": "^16.0.6",
     "style-loader": "^4.0.0",
